Interface file moved to dedicated directory
