+++ Gems/SimulationInterfacesROS2/Code/Source/Services/GetEntityStateServiceHandler.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "GetEntityStateServiceHandler.h"
-#include "AzCore/std/optional.h"
+#include <AzCore/std/optional.h>
 #include <ROS2/ROS2Bus.h>
 #include <ROS2/Utilities/ROS2Conversions.h>
 #include <SimulationInterfaces/SimulationEntityManagerRequestBus.h>
