+++ Gems/SimulationInterfacesROS2/Code/Source/Services/SetEntityStateServiceHandler.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "SetEntityStateServiceHandler.h"
-#include "AzCore/std/optional.h"
+#include <AzCore/std/optional.h>
 #include <ROS2/Utilities/ROS2Conversions.h>
 #include <SimulationInterfaces/SimulationEntityManagerRequestBus.h>
